Change config files into a separate file
